requirements.txt
pandas
requests
flask
plotly
matplotlib
scikit-learn
tensorflow
vaderSentiment
python-dotenv

